bot/requirements.txt
APScheduler==3.11.0
fastapi==0.114.1
python-telegram-bot==21.7
rich==13.9.4
SQLAlchemy==2.0.36
uvicorn==0.34.0
